Consolidate trait extensions
